requirements.txt
pydantic==1.10.2
PyYAML==6.0
